Update xp-pen-tablet from 3.2.3.230215_1 to 3.4.9.231023_1
License was changed to custom:Propietary as it now includes an EULA which explicitly indicates that redistribution and/or modifying is not permitted. The LGPL license file on previous versions was only referring to the bundled Qt5 libraries, not the software itself. This was made clear with the inclusion of ThirdPartyLibraries.html file.

In consequence, any package of this software in cereus-extra's mirrors will be removed ASAP. Repository was changed to nonfree in the template (this is the first propietary package to be included here, so it is unnecesary to use 'cereus-nonfree').

The glibc containerized package for musl (xp-pen-tablet-musl) will be temporarily removed as the current install method does not longer work (it downloads the package from cereus-extra/x86_64) so it has to be redo.